08.12 - final refactorings (step 3 - refactor the products feature) - create a products package. move all the product related code to the product package
