+++ src/main/java/com/jparkkennaby/store/entities/CartItem.java
@@ -2,6 +2,8 @@ package com.jparkkennaby.store.entities;
 
 import java.math.BigDecimal;
 
+import com.jparkkennaby.store.products.Product;
+
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
