+++ src/main/java/com/jparkkennaby/store/products/CategoryRepository.java
@@ -1,8 +1,6 @@
-package com.jparkkennaby.store.repositories;
+package com.jparkkennaby.store.products;
 
 import org.springframework.data.repository.CrudRepository;
-
-import com.jparkkennaby.store.entities.Category;
 
 public interface CategoryRepository extends CrudRepository<Category, Byte> {
 }
