+++ src/main/java/com/jparkkennaby/store/products/ProductMapper.java
@@ -1,11 +1,8 @@
-package com.jparkkennaby.store.mappers;
+package com.jparkkennaby.store.products;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
-
-import com.jparkkennaby.store.dtos.ProductDto;
-import com.jparkkennaby.store.entities.Product;
 
 @Mapper(componentModel = "spring") // so spring is made aware and can make beans at runtime
 public interface ProductMapper {
