+++ src/main/java/com/jparkkennaby/store/products/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.exceptions;
+package com.jparkkennaby.store.products;
 
 public class ProductNotFoundException extends RuntimeException {
