+++ src/main/java/com/jparkkennaby/store/users/User.java
@@ -8,7 +8,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import com.jparkkennaby.store.entities.Product;
+import com.jparkkennaby.store.products.Product;
 
 @Setter
 @Getter
